requirements.txt
streamlit
tensorflow==2.20.0
opencv-python-headless
numpy
pillow